Extract TestingDirectTrinoClient
